Corrección en decimales

+++ database/migrations/2020_12_29_170928_create_lotes_table.php
@@ -31,7 +31,7 @@ class CreateLotesTable extends Migration
             $table->string('estado_egresos', 1);
             $table->string('observaciones')->nullable();
             $table->integer('cant_ahogados')->nullable();
-            $table->integer('peso_ahogados')->nullable();
+            $table->decimal('peso_ahogados')->nullable();
             $table->timestamps();
         });
     }
